XWIKI-15065: Replace Silk icons with Icon Themes values in attachments Upload component (#2700)
* Added icons to the generated node.

Note: I decided to go with a CSS only logic: instead of creating and deleting the icons based on the status, we use the logic already in place to hide/show the icons (all created at the start). There's only 4 of them, and they are not generated on every page load, so I think this is an appropriate solution.