XWIKI-12987: Relative links are made absolute or even broken after moving a page
WIP
The idea of this work is to:
  1. Provide a way to access all documents that are moved as part of a
     move job
  2. Use that information when performing a call to ReferenceRenamer to
     define if a relative untyped link should be handled or not

On top of it, the idea is also to check if the doc exists in case of
refactoring of a link to avoid refactoring unexisting relative links.
One problem is remaining about relative link pointing to sibling pages
(e.g. the link to Alice in Bob page in the ticket): we rely apparently
to an old mechanism for backward compatibility reason for this to work
in the UI, we might need same thing in the check, or to decide to
ignore that UC.

I started to add an integration tests but for some reason it's not
passing, though it seemed to be working locally for the scenario
described in the ticket (except for the link in Bob page).