XWIKI-22978: Comments are not ordered by date
Provide a quickfix for ordering the comments directly in
commentsinline.vm. Since we cannot call directly $collectiontool.sort
because the Object property cannot be accessed by beanutils, we build a
map indexed by date and we perform the ordering manually, taking into
account the rare possibility that 2 comments have been added at exact
same date.
Also all this work is duplicated in comments2.vm for fixing it in old
PDF export.
Note that ideally all this work should be refactored with a proper Java
API for comments, see also XWIKI-23049.