XWIKI-7902: Remove dependency on retired commons-httpclient
- Replace commons-httpclient:commons-httpclient with org.apache.httpcomponents:httpclient
- Add enforcer rule to prevent future use of commons-httpclient
- Update InstallMojo to use new httpcomponents API
- Update all pom.xml files to use new dependency

This addresses the JIRA issue XWIKI-7902 by removing the dependency on the retired
commons-httpclient library and replacing it with the modern org.apache.httpcomponents:httpclient.
The API has changed significantly, so code using the old API needs to be rewritten.
An enforcer rule has been added to prevent any future modules from using the old dependency.